k8s_project_dockerimg/docker_img/sang_update_model/requirements.txt
kubeflow-katib
PyYAML==5.2
scikit-learn==0.24.2
pandas==1.1.5
joblib==1.1.1
xgboost==1.5.2
